hash table

+++ binary search/367 ValidPerfectSquare.cpp
@@ -11,10 +11,10 @@
 class Solution {
 public:
     bool isPerfectSquare(int num) {
-        int left = 1;
-        int right = num;
+        long left = 1;
+        long right = num;
         while (left <= right) {
-            long long mid = left + (right - left) / 2;
+            long  mid = left + (right - left) / 2;
             if (mid * mid == num) {
                 return true;
             } else if(mid * mid < num) {
